+++ src/app/ev/[id]/loading.tsx
@@ -1,9 +1,9 @@
 import React from 'react'
 
-const loading = () => {
+const Loading = () => {
   return (
     <div>Loading...</div>
   )
 }
 
-export default loading
+export default Loading
