+++ src/components/ui/ImageGallery.tsx
